+++ meerket/meerket-application/src/main/java/org/j1p5/api/product/controller/ProductController.java
@@ -1,12 +1,11 @@
 package org.j1p5.api.product.controller;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.j1p5.api.global.annotation.LoginUser;
 import org.j1p5.api.global.response.Response;
 import org.j1p5.api.product.converter.MultipartFileConverter;
-import org.j1p5.api.product.dto.request.ProductRequestDto;
+import org.j1p5.api.product.dto.request.ProductCreateRequestDto;
 import org.j1p5.api.product.dto.request.ProductUpdateRequest;
 import org.j1p5.common.annotation.CursorDefault;
 import org.j1p5.common.dto.Cursor;
@@ -15,7 +14,6 @@ import org.j1p5.domain.product.dto.ProductInfo;
 import org.j1p5.domain.product.dto.ProductResponseDetailInfo;
 import org.j1p5.domain.product.dto.ProductResponseInfo;
 import org.j1p5.domain.product.service.ProductService;
-import org.j1p5.domain.user.entity.UserEntity;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -45,7 +43,7 @@ public class ProductController {
      * @author sunghyun0610
      */
     @PostMapping(consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
-    public Response<Void> makeProduct(@RequestPart(name = "request") ProductRequestDto request,
+    public Response<Void> makeProduct(@RequestPart(name = "request") ProductCreateRequestDto request,
                                 @RequestPart(name = "images", required = false) List<MultipartFile> images,
                                 @LoginUser Long userId
     ) {
@@ -54,7 +52,7 @@ public class ProductController {
         // 세션이들어옴 -> 세션에있는 userid뽑아냄 ->이거 이용해서 지역인증 테이블에서 findById , 추후에 user role추가
 
 
-        ProductInfo productInfo = ProductRequestDto.toInfo(request);
+        ProductInfo productInfo = ProductCreateRequestDto.toInfo(request);
         List<File> imageFiles = new ArrayList<>();
 
         if (images != null && !images.isEmpty()) {
